Fix site names breaking on 'test'
Although 'test' is a reserved environment name, it's being used in many
sites that are getting frozen. Since the current RegEx treats 'test' as
an environment name, it strips that out and breaks the command from
thawing out sites like 'test-my-cool-feature'.

This dumbs down the check a bit and only looks for 'dev' as the
environment name, since it's the most likely env to get passed (live
sites don't get frozen). It also splits the replacement into 3 calls
for ease of reading and debugging until a better RegEx comes along.